% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_classes.R
\name{clean_classes}
\alias{clean_classes}
\title{Clean Classes}
\usage{
clean_classes(D, assignment, classes = "all", tau = c(0.01, 0.05, 0.1),
  labels = NULL, display_progress = FALSE)
}
\arguments{
\item{D}{A distance matrix containing the pairwise dissimilarity scores between instances}

\item{assignment}{The assigned group of each instance}

\item{classes}{The subset of classes on which filtering is performed, or "all" if all classes should be analyzed.}

\item{labels}{identifier for each instance.  If NULL, the row/column names or indices of D are used}
}
\description{
Test whether each instance in a class actually belongs.
}
\details{
For each instance in an analyzed class, this function will estimate the probability that it was correctly placed in that class.
}
