% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data
A convience wrapper around simulate_clustered_data for generating clustered data sets}
\usage{
generate_data(
  Nk,
  p_mismatch,
  p_garbage,
  K_proteome,
  n,
  s = 1,
  tau = 1,
  rho = 0
)
}
\arguments{
\item{Nk}{The number of instances in each class (can be a table)}

\item{p_mismatch}{The expected proprotion of instances in any given class which are misclassified}

\item{p_garbage}{The expected proportion of instances in any given class which are actually quantitative noise.}

\item{n}{The number of samples}

\item{s}{The within-class standard deviation}

\item{tau}{The between-class standard deviation}

\item{rho}{The average correlation between different classes (must be greater than -1 / K, where K is the number of classes)}
}
\description{
Generate data
A convience wrapper around simulate_clustered_data for generating clustered data sets
}
