% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{identify_outliers}
\alias{identify_outliers}
\title{Identify Outliers}
\usage{
identify_outliers(assignment, D, B, min_group_size, labels,
  display_progress = FALSE)
}
\arguments{
\item{assignment}{A character vector containing the group to which each entity has been classified}

\item{D}{A numeric vector containing the distance or quasi-distance between the entities.}

\item{B}{The number of simulated draws to use in determining the distribution of the means.}

\item{min_group_size}{The smallest number of entities required to perform filtering.  (Must be at least 2)}

\item{labels}{entity-specific labels.}
}
\description{
Calculate the probability that each entity in a group with at least 2 entities actually belongs
}
\details{
For each entity in group k, this function estimates the probability the mean distance between it and another entity in group k has the same distribution as a mean distance from it and another group.
Under the null hypothesis, the mean distance is no different, and we remove it from the group.
Under the alterantive hypothesis, the mean distance is closer, and we conclude that the entity truly belongs to the group.
}
