% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_clustered_data.R
\name{simulate_clustered_data}
\alias{simulate_clustered_data}
\title{Simulate clustered data using the normal distribution.}
\usage{
simulate_clustered_data(n = 100, Nk = c(40, 200), s = c(1, 1),
  rho = matrix(c(0.6, 0.1, 0.1, 0.25), nrow = 2, ncol = 2), tau = 1,
  method = c("by-class", "by-instance"))
}
\arguments{
\item{n}{The total number of observations per instance.}

\item{Nk}{A vector of length \eqn{K} giving the number of  within each class}

\item{s}{A vector of standard deviations. See details.}

\item{rho}{A matrix of correlation coefficients.  See details.}

\item{tau}{The within-group variance.  Only used when method = "by_class".}

\item{method}{Either "by_instance" or "by_class".}
}
\description{
This function generates a matrix with \eqn{n} indepenent rows and \eqn{N} columns,
where the columns are clustered into \eqn{K} classes with \eqn{N_k}{N[k]} instances in class \eqn{k}
}
\details{
This function generates a matrix with a block - correlation stucture across columns and independent rows.
When method = 'by_instance', the values of s and rho are taken to be instance-level properties of the data.
That is, s is a vector of length K such that the ith entry is the standard deviations of observations within class k
and rho is a  \eqn{K \times K}{K * K} symmetric matrix such that entry (i,j) gives the correlation between an instance is class i and an instance in class j.  Correspondingly, entry (i,i) gives the correlation between two (different) instances in class i.
In contrast, when method = 'by_class', the values of s and rho are taken to be class-level properties.
The variance from the 'by_instance' characterization is broken down into a class-level variance (s^2), which gives the variability of the "true" pattern of the class over the observations"  and an instance-level variance (tau^2).which gives the variabilty of the the observed instances from the true pattern. The correlation is now in terms of the classes: entry (i,j) gives the correlation between the "true" pattern of class i and the "true" pattern of class j.
}
\examples{
rho.mat <- matrix(c(.5,.2,.2,.3),nrow = 2,ncol = 2)
X <- simulate_by_instance(Nk = c(50, 100),rho = rho.mat,n = 150)
}
