% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test}
\alias{test}
\title{test function}
\usage{
test(vec1)
}
\arguments{
\item{vec1}{a numeric vector}
}
\description{
this is a trivial example.
}
\details{
a trivial example to figure out why matrix multiplication is giving errors.
}
